AV following waypoint lane
A tesla roadster now follows the waypoint lane. It has a duplicate script so it can be modified to encompany also delays or speed changes and simplifies the making of the lane change.

The waypoints themselves have been transformed so that they are stuck to the ground. This has effect on the truck.

Changing the waypoint system to a bezier curve could improve the realism later on.