Update truck and AV
Both truck and AV have updated rigid bodies and wheel colliders to run more smoothly and be able to be hit. This also meant that the truck had to start one waypoint earlier, which is accomplished in the script.

Also the placeholder cube has been removed.

This commit is reverting the project back to windows, which will be rectified in the upcoming merge.